Add private and public handlers file. Fix tests
